# openapi.yaml
openapi: 3.0.3
info:
  title: API NetWish
  description: API for recommendation films and series, based on the user's preferences in TMDB.
  contact:
    email: dev@netWish.com
  version: 2.0.0
servers:
  - url: http://localhost:3000
    description: Local server
    variables:
      protocol:
        default: http
        enum:
          - http
          - https
      server:
        default: localhost
...
paths:
...
  /films:
    get:
      tags:
        - films
      summary: Get all films
      description: Get all films
      operationId: getFilms
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/films'
        '400':
          description: Bad request
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Not found
        '500':
          description: Internal server error
...
components:
  schemas:
...
    films:
      type: object
      properties:
        page:
          type: integer
        result:
          type: object
          properties:
            poster_path:
              type: string
            adult:
              type: boolean
            overview:
              type: string
            release_date:
              type: string
            genre_ids:
              type: array
              items:
                type: integer
            id:
              type: integer
            original_title:
              type: string
            original_language:
              type: string
            title:
              type: string
            backdrop_path:
              type: string
            popularity:
              type: number
            vote_count:
              type: number
            video:
              type: boolean
            vote_average:
              type: number
...